Naughty RT-Thread fiddles around with private members of newlib.
_GLOBAL_REENT->__sdidinit is defined in older versions of newlib bundled with the standard arm-none-eabi toolchains published by ARM.
It seems to be present in gcc-arm-none-eabi-10.3-2021.07 and earlier, but has gone by arm-gnu-toolchain-12.3.rel1-x86_64-arm-none-eabi

I grepped the rt code and the toolchains for references.
It's wrapped in the _REENT_SMALL_CHECK_INIT macro in sys/reent.h
But this has been redefined to an empty expansion in later newlib versions.

rt doesn't appear to use this anywhere either directly or via the macro.
So I simply commented out the assignment.