A hack to get the build going..
rtconfig.py specifies toolchain locations and options.
It's a bit of a disaster since it hard codes paths to the toolchain binaries.

I've added a symlink at the repo root that points to one of the arm-none-eabi installations on my system,
and made EXEC_PATH in rtconfig.py point to that symlink.

The build is a bit of a disaster.. but I just want to get things working so that I can port the W60x drivers to another OS.