Fix build for most examples.
All the examples build except for:

  24_iot_websocket
  34_iot_cloud_ms_azure
  35_iot_cloud_tencent

24 and 35 have the same problem,
namely include conflicts between the eabi toolchain implementation of pthreads and rt's implementation.

34 has a problem with include paths.
The azure source code contains a file called 'strings.h'
This conflicts with the standard C library, which has a header of the same name.

Generally things are ok, because the azure strings.h is included by the azure source via a path, namely
#include "azure_c_shared_utility/strings.h"

But the macro_utils.h header happens to include <string.h>,
which then includes <strings.h> but instead of the standard library version it resolves to the azure version.

A hack-around was proposed here:

https://github.com/Azure/azure-iot-sdk-c/issues/1177#issuecomment-535581937

Basically do an exhaustive name change for azure's strings.h to az_strings.h
This touches a lot of files in the azure source tree.
I might try this in a subsequent commit...